+++ app/backend/install.sh
@@ -1,4 +1,0 @@
-yarn install
-
-yarn migrate:dev
-yarn codegen:prisma