+++ app/backend/prisma/migrations/20220423164543_init/migration.sql
@@ -1,8 +1,0 @@
--- CreateTable
-CREATE TABLE "message" (
-    "id" SERIAL NOT NULL,
-    "created_at" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    "content" TEXT NOT NULL,
-
-    CONSTRAINT "message_pkey" PRIMARY KEY ("id")
-);