+++ clients/redis/js/src/MemorixBaseApi.ts
@@ -1,11 +1,53 @@
 import Redis from "ioredis";
-import { CacheItem, PubsubItem, TaskItem } from "./types";
+import { v4 as uuidv4 } from "uuid";
+import callbackToAsyncIterator from "callback-to-async-iterator";
+import {
+  CacheItem,
+  CacheSetOptions,
+  TaskDequequeOptions,
+  PubsubCallback,
+  PubsubItem,
+  TaskItem,
+  CacheItemNoKey,
+  PubsubItemNoKey,
+  TaskItemNoKey,
+} from "./types";
 import { hashKey } from "./utils/hashKey";
 
-export class MemorixBaseApi {
-  private readonly redis = new Redis(process.env.REDIS_ENV!);
+type Defaults = {
+  cacheSetOptions?: CacheSetOptions;
+  taskDequequeOptions?: TaskDequequeOptions;
+};
 
-  private readonly redisSub = new Redis(process.env.REDIS_ENV!);
+export class MemorixBaseApi {
+  private readonly redis: Redis;
+
+  private readonly redisSub: Redis;
+
+  private readonly defaults?: Defaults;
+
+  private redisTasks: Redis[];
+
+  constructor({
+    redisUrl,
+    defaults,
+  }: {
+    redisUrl: string;
+    defaults?: Defaults;
+  }) {
+    this.redis = new Redis(redisUrl);
+    this.redisSub = this.redis.duplicate();
+    this.redisTasks = [];
+    this.defaults = defaults;
+  }
+
+  disconnect(): void {
+    this.redis.disconnect();
+    this.redisSub.disconnect();
+    this.redisTasks.forEach((x) => {
+      x.disconnect();
+    });
+  }
 
   getCacheItem<Key, Payload>(identifier: string): CacheItem<Key, Payload> {
     const hashCacheKey = (key: Key | undefined) => {
@@ -13,13 +55,22 @@ export class MemorixBaseApi {
     };
 
     return {
-      set: async (...args) => {
-        const key = args.length === 1 ? undefined : args[0];
-        const payload = args.length === 1 ? args[0] : args[1];
+      set: async (key, payload, options) => {
+        const { expire = undefined } = {
+          ...this.defaults?.cacheSetOptions,
+          ...options,
+        };
         const hashedKey = hashCacheKey(key);
-        await this.redis.set(hashedKey, JSON.stringify(payload));
+        const params = expire
+          ? [expire.isInMs ? "PX" : "EX", expire.value.toString()]
+          : [];
+        await this.redis.set(
+          hashedKey,
+          JSON.stringify(payload),
+          ...(params as any)
+        );
       },
-      get: async (...[key]) => {
+      get: async (key) => {
         const hashedKey = hashCacheKey(key);
         const found = await this.redis.get(hashedKey);
         if (found) {
@@ -30,115 +81,191 @@ export class MemorixBaseApi {
     };
   }
 
+  getCacheItemNoKey<Payload>(...itemArgs: any[]): CacheItemNoKey<Payload> {
+    const item = (this.getCacheItem as any)(...itemArgs);
+
+    return {
+      get: (...args) => item.get(undefined, ...args),
+      set: (...args) => item.set(undefined, ...args),
+    };
+  }
+
   getPubsubItem<Key, Payload>(identifier: string): PubsubItem<Key, Payload> {
     const hashPubsubKey = (key: Key | undefined) => {
       return hashKey(key ? [identifier, key] : [identifier]);
     };
 
     return {
-      publish: async (...args) => {
-        const key = args.length === 1 ? undefined : args[0];
-        const payload = args.length === 1 ? args[0] : args[1];
+      publish: async (key, payload) => {
         const hashedKey = hashPubsubKey(key);
-        await this.redis.publish(hashedKey, JSON.stringify(payload));
+        const subscribersSize = await this.redis.publish(
+          hashedKey,
+          JSON.stringify(payload)
+        );
+        return { subscribersSize };
       },
-      subscribe: async (...args) => {
-        const key = args.length === 1 ? undefined : args[0];
-        const callback = args.length === 1 ? args[0] : args[1];
+      subscribe: ((
+        key: Key,
+        callback: PubsubCallback<{ payload: Payload }> | undefined
+      ) => {
         const hashedKey = hashPubsubKey(key);
 
-        return new Promise((resolve, reject) => {
-          this.redisSub.subscribe(hashedKey, (err, count) => {
-            if (err) {
-              console.error("Failed to subscribe: %s", err.message);
-              reject(err);
-            } else {
-              console.log(
-                `Subscribed successfully! This client is currently subscribed to ${count} channels.`
-              );
-              resolve();
-            }
+        const getListenPromise = (cb: NonNullable<typeof callback>) =>
+          new Promise<{ stop(): Promise<void> }>((resolve, reject) => {
+            this.redisSub.subscribe(hashedKey, (err) => {
+              if (err) {
+                reject(err);
+              } else {
+                resolve({
+                  stop: async () => {
+                    await this.redisSub.unsubscribe(hashedKey);
+                  },
+                });
+              }
+            });
+
+            this.redisSub.on("message", (group, payload) => {
+              if (hashedKey === group) {
+                cb({ payload: JSON.parse(payload) });
+              }
+            });
           });
 
-          this.redisSub.on("message", (group, payload) => {
-            if (hashedKey === group) {
-              // console.log(`got payload ${payload} in key ${group}`);
-              callback(JSON.parse(payload));
-            }
+        if (callback === undefined) {
+          return callbackToAsyncIterator<
+            { payload: Payload },
+            { stop: () => Promise<void> }
+          >((cb) => getListenPromise(cb), {
+            onClose({ stop }) {
+              stop();
+            },
           });
-        });
-      },
+        }
+
+        return getListenPromise(callback);
+      }) as any,
+    };
+  }
+
+  getPubsubItemNoKey<Payload>(...itemArgs: any[]): PubsubItemNoKey<Payload> {
+    const item = (this.getPubsubItem as any)(...itemArgs);
+
+    return {
+      publish: (...args) => item.publish(undefined, ...args),
+      subscribe: (...args) => (item.subscribe as any)(undefined, ...args),
     };
   }
 
   getTaskItem<Key, Payload, Returns>(
-    identifier: string
+    identifier: string,
+    hasReturns: Returns extends undefined ? false : true
   ): TaskItem<Key, Payload, Returns> {
     const hashPubsubKey = (key: Key | undefined) => {
       return hashKey(key ? [identifier, key] : [identifier]);
     };
 
+    const returnTask = hasReturns
+      ? this.getTaskItem<string, Returns, undefined>(
+          `${identifier}_returns`,
+          false
+        )
+      : undefined;
+
     return {
-      queue: async (...args) => {
-        const key = args.length === 1 ? undefined : args[0];
-        const payload = args.length === 1 ? args[0] : args[1];
+      queue: async (key, payload) => {
         const hashedKey = hashPubsubKey(key);
+        const returnsId = hasReturns ? uuidv4() : undefined;
 
-        await this.redis.publish(hashedKey, JSON.stringify(payload));
+        const queueSize = await this.redis.rpush(
+          hashedKey,
+          JSON.stringify(hasReturns ? [returnsId, payload] : [payload])
+        );
 
-        return new Promise((resolve, reject) => {
-          this.redisSub.subscribe(`returns_${hashedKey}`, (err, count) => {
-            if (err) {
-              console.error("Failed to subscribe: %s", err.message);
-              reject(err);
-            } else {
-              console.log(
-                `Subscribed successfully! This client is currently subscribed to ${count} channels.`
-              );
-            }
-          });
+        const returnsPromise = hasReturns
+          ? new Promise((res, rej) => {
+              let stop: () => Promise<void> | undefined;
+              returnTask!
+                .dequeue(returnsId!, ({ payload: returns }) => {
+                  stop();
+                  res(returns);
+                })
+                .then((dequeueObj) => {
+                  stop = dequeueObj.stop;
+                })
+                .catch(rej);
+            })
+          : undefined;
 
-          this.redisSub.on("message", (group, returnedPayload) => {
-            if (hashedKey === group) {
-              // console.log(`got payload ${payload} in key ${group}`);
-              this.redisSub.unsubscribe(`returns_${hashedKey}`);
-
-              resolve(JSON.parse(returnedPayload));
-            }
-          });
-        });
+        if (hasReturns) {
+          return {
+            queueSize,
+            getReturns: () => returnsPromise,
+          };
+        }
+        return {
+          queueSize,
+        } as any;
       },
-      dequeue: async (...args) => {
-        const key = args.length === 1 ? undefined : args[0];
-        const callback = args.length === 1 ? args[0] : args[1];
+      dequeue: async (key, callback, options) => {
         const hashedKey = hashPubsubKey(key);
+        const { takeNewest = false } = {
+          ...this.defaults?.taskDequequeOptions,
+          ...options,
+        };
 
-        return new Promise((resolve, reject) => {
-          this.redisSub.subscribe(hashedKey, (err, count) => {
+        const redisClient = this.redis.duplicate();
+        this.redisTasks.push(redisClient);
+
+        const stoppedPromise = new Promise((res) => {
+          const cb = async (err: any, blpop: any) => {
             if (err) {
-              console.error("Failed to subscribe: %s", err.message);
-              reject(err);
-            } else {
-              console.log(
-                `Subscribed successfully! This client is currently subscribed to ${count} channels.`
-              );
-              resolve();
+              res(err);
             }
-          });
+            if (blpop) {
+              redisClient[takeNewest ? "brpop" : "blpop"](hashedKey, 0, cb);
 
-          this.redisSub.on("message", (group, payload) => {
-            if (hashedKey === group) {
-              // console.log(`got payload ${payload} in key ${group}`);
-              const returnedPayload = callback(JSON.parse(payload));
+              const [, wrapedPayloadStr] = blpop;
+              const wrapedPayload = JSON.parse(wrapedPayloadStr);
+              const payload: Payload = hasReturns
+                ? wrapedPayload[1]
+                : wrapedPayload[0];
 
-              this.redis.publish(
-                `returns_${hashedKey}`,
-                JSON.stringify(returnedPayload)
-              );
+              const result = callback({ payload });
+              if (hasReturns) {
+                const returnsId: string = wrapedPayload[0];
+                const returns = (
+                  result instanceof Promise ? await result : result
+                ) as Returns;
+                await returnTask!.queue(returnsId, returns);
+              }
             }
-          });
+          };
+
+          redisClient[takeNewest ? "brpop" : "blpop"](hashedKey, 0, cb);
         });
+
+        return {
+          stop: async () => {
+            redisClient.disconnect();
+            const indexToRemove = this.redisTasks.indexOf(redisClient);
+            if (indexToRemove !== -1) {
+              this.redisTasks.splice(indexToRemove, 1);
+            }
+            await stoppedPromise;
+          },
+        };
       },
+    };
+  }
+
+  getTaskItemNoKey<Payload, Returns>(
+    ...itemArgs: any[]
+  ): TaskItemNoKey<Payload, Returns> {
+    const item = (this.getTaskItem as any)(...itemArgs);
+
+    return {
+      queue: (...args) => item.queue(undefined, ...args),
+      dequeue: (...args) => item.dequeue(undefined, ...args),
     };
   }
 }