+++ clients/redis/js/src/example-schema.test.ts
@@ -1,0 +1,163 @@
+import { MemorixApi, Animal } from "./example-schema.generated";
+
+const redisUrl = process.env.REDIS_URL!;
+
+describe("example schema has", () => {
+  let memorixApi: MemorixApi;
+  beforeEach(() => {
+    memorixApi = new MemorixApi({ redisUrl });
+  });
+  afterEach(() => {
+    memorixApi.disconnect();
+  });
+  it("cache", async () => {
+    await memorixApi.cache.user.set("uv", { name: "uv", age: 29 });
+    const user = await memorixApi.cache.user.get("uv");
+    expect(user!.age).toBe(29);
+  });
+  it("cache expire", async () => {
+    await memorixApi.cache.user.set(
+      "uv",
+      { name: "uv", age: 29 },
+      { expire: { value: 1 } }
+    );
+    const user1 = await memorixApi.cache.user.get("uv");
+    expect(user1!.age).toBe(29);
+    await new Promise((res) => setTimeout(res, 1500));
+    const user2 = await memorixApi.cache.user.get("uv");
+    expect(user2).toBe(null);
+  });
+  describe("pubsub", () => {
+    it("publish says how many subscribers", (done) => {
+      memorixApi.pubsub.message
+        .subscribe(() => {})
+        .then(() => {
+          memorixApi.pubsub.message
+            .publish("hello uv")
+            .then(({ subscribersSize }) => {
+              try {
+                expect(subscribersSize).toBe(1);
+                done();
+              } catch (error) {
+                done(error);
+              }
+            });
+        });
+    });
+    it("subscribe gets payload", (done) => {
+      memorixApi.pubsub.message
+        .subscribe(({ payload }) => {
+          try {
+            expect(payload).toBe("hello uv");
+            done();
+          } catch (error) {
+            done(error);
+          }
+        })
+        .then(() => {
+          memorixApi.pubsub.message.publish("hello uv");
+        });
+    });
+    it("subscribe gets payload with AsyncIterableIterator", async () => {
+      setTimeout(() => {
+        memorixApi.pubsub.message.publish("hello uv");
+      }, 500);
+      for await (const { payload } of memorixApi.pubsub.message.subscribe()) {
+        expect(payload).toBe("hello uv");
+        break;
+      }
+    });
+  });
+  describe("task", () => {
+    beforeEach(async () => {
+      const { stop } = await memorixApi.task.runAlgo.dequeue(() => {
+        return Animal.dog;
+      });
+      await new Promise((res) => setTimeout(res, 1000));
+      await stop();
+    });
+    it("queue returns the queue size", async () => {
+      await memorixApi.task.runAlgo.queue("uv1");
+      const { queueSize } = await memorixApi.task.runAlgo.queue("uv2");
+      expect(queueSize).toBe(2);
+    });
+    it("dequeue receives a message", (done) => {
+      memorixApi.task.runAlgo
+        .queue("uv3")
+        .then(() => memorixApi.task.runAlgo.queue("uv4"))
+        .then(({ queueSize }) => {
+          expect(queueSize).toBe(2);
+
+          return new Promise((res) => {
+            const payloads: string[] = [];
+            let stop: any;
+            memorixApi.task.runAlgo
+              .dequeue(({ payload }) => {
+                payloads.push(payload);
+                if (payloads.length === 2) {
+                  res({
+                    payloads,
+                    stop,
+                  });
+                }
+                return Animal.cat;
+              })
+              .then((args) => {
+                stop = args.stop;
+              });
+          });
+        })
+        .then(async ({ payloads, stop }: any) => {
+          expect(payloads).toStrictEqual(["uv3", "uv4"]);
+          await stop();
+          done();
+        })
+        .catch(done);
+    });
+    it("dequeue receives a message in opposite order if asked", (done) => {
+      memorixApi.task.runAlgo
+        .queue("uv6")
+        .then(() => memorixApi.task.runAlgo.queue("uv7"))
+        .then(({ queueSize }) => {
+          expect(queueSize).toBe(2);
+
+          return new Promise((res) => {
+            const payloads: string[] = [];
+            let stop: any;
+            memorixApi.task.runAlgo
+              .dequeue(
+                ({ payload }) => {
+                  payloads.push(payload);
+                  if (payloads.length === 2) {
+                    res({
+                      payloads,
+                      stop,
+                    });
+                  }
+                  return Animal.cat;
+                },
+                { takeNewest: true }
+              )
+              .then((args) => {
+                stop = args.stop;
+              });
+          });
+        })
+        .then(async ({ payloads, stop }: any) => {
+          expect(payloads).toStrictEqual(["uv7", "uv6"]);
+          await stop();
+          done();
+        })
+        .catch(done);
+    });
+    it("queue receives a returns", async () => {
+      const { stop } = await memorixApi.task.runAlgo.dequeue(({ payload }) =>
+        payload === "uv5" ? Animal.person : Animal.dog
+      );
+      const { getReturns } = await memorixApi.task.runAlgo.queue("uv5");
+      const returns = await getReturns();
+      expect(returns).toBe(Animal.person);
+      await stop();
+    });
+  });
+});