+++ clients/redis/python/memorix_client_redis/__init__.py
@@ -1,8 +1,15 @@
-from .features.api.api import Api as MemorixClientApi
+from .features.api.api import (
+    Api as MemorixClientApi,
+    ApiDefaults as MemorixClientApiDefaults,
+)
 from .features.api.cache.cache_api import CacheApi as MemorixClientCacheApi
 from .features.api.cache.cache_item import (
     CacheItem as MemorixClientCacheApiItem,
     CacheItemNoKey as MemorixClientCacheApiItemNoKey,
+)
+from .features.api.cache.cache_options import (
+    CacheSetOptions as MemorixClientCacheSetOptions,
+    CacheSetOptionsExpire as MemorixClientCacheSetOptionsExpire,
 )
 from .features.api.pubsub.pubsub_api import PubSubApi as MemorixClientPubSubApi
 from .features.api.pubsub.pubsub_item import (
@@ -15,5 +22,8 @@ from .features.api.task.task_item import (
     TaskItemNoKey as MemorixClientTaskApiItemNoKey,
     TaskItemNoReturns as MemorixClientTaskApiItemNoReturns,
     TaskItemNoKeyNoReturns as MemorixClientTaskApiItemNoKeyNoReturns,
+)
+from .features.api.task.task_options import (
+    TaskDequequeOptions as MemorixClientTaskDequequeOptions,
 )
 from dataclasses import dataclass