+++ clients/redis/python/memorix_client_redis/example_schema_generated.py
@@ -1,23 +1,31 @@
 import typing
 from enum import Enum
-from memorix_client_redis import (
+from memorix_client_redis import (  # noqa: F401
     dataclass,
     MemorixClientApi,
     MemorixClientCacheApi,
     MemorixClientCacheApiItem,
-    # MemorixClientCacheApiItemItemNoKey,
+    MemorixClientCacheApiItemNoKey,
+    MemorixClientCacheSetOptions as _MemorixClientCacheSetOptions,
+    MemorixClientCacheSetOptionsExpire as _MemorixClientCacheSetOptionsExpire,
     MemorixClientPubSubApi,
-    # MemorixClientPubSubApiItem,
+    MemorixClientPubSubApiItem,
     MemorixClientPubSubApiItemNoKey,
     MemorixClientTaskApi,
-    # MemorixClientTaskApiItem,
+    MemorixClientTaskApiItem,
     MemorixClientTaskApiItemNoKey,
-    # MemorixClientTaskApiItemNoReturn,
-    # MemorixClientTaskApiItemNoKeyNoReturns
+    MemorixClientTaskApiItemNoReturns,
+    MemorixClientTaskApiItemNoKeyNoReturns,
+    MemorixClientTaskDequequeOptions as _MemorixClientTaskDequequeOptions,
 )
 
 
-class Animal(Enum):
+MemorixClientCacheSetOptions = _MemorixClientCacheSetOptions
+MemorixClientCacheSetOptionsExpire = _MemorixClientCacheSetOptionsExpire
+MemorixClientTaskDequequeOptions = _MemorixClientTaskDequequeOptions
+
+
+class Animal(str, Enum):
     dog = "dog"
     cat = "cat"
     person = "person"