+++ clients/redis/python/memorix_client_redis/features/api/task/task_options.py
@@ -1,0 +1,9 @@
+from typing import Optional
+
+
+class TaskDequequeOptions(object):
+    def __init__(
+        self,
+        take_newest: Optional[bool] = False,
+    ) -> None:
+        self.take_newest = take_newest