+++ CHANGELOG.md
@@ -1,3 +1,37 @@
+# [1.0.0-rc.20](https://github.com/uvop/memorix/compare/v1.0.0-rc.19...v1.0.0-rc.20) (2022-09-23)
+
+
+### Bug Fixes
+
+* Some docs improvements ([18af5ac](https://github.com/uvop/memorix/commit/18af5acea629732155f47262bf0a15cb73773bab))
+
+# [1.0.0-rc.19](https://github.com/uvop/memorix/compare/v1.0.0-rc.18...v1.0.0-rc.19) (2022-09-21)
+
+
+### Bug Fixes
+
+* try deploy docs ([c442ed5](https://github.com/uvop/memorix/commit/c442ed5f2533762e4659d1dab3ce35748bb54243))
+
+# [1.0.0-rc.18](https://github.com/uvop/memorix/compare/v1.0.0-rc.17...v1.0.0-rc.18) (2022-09-21)
+
+
+### Bug Fixes
+
+* try publish again ([e51593c](https://github.com/uvop/memorix/commit/e51593cd322c03de7e8f2d89578d9b3ddf494d5d))
+
+# [1.0.0-rc.17](https://github.com/uvop/memorix/compare/v1.0.0-rc.16...v1.0.0-rc.17) (2022-09-21)
+
+
+### Bug Fixes
+
+* build client js ([2dbba3f](https://github.com/uvop/memorix/commit/2dbba3f46a1918d489bb292345b0c867cf754b7f))
+* docs publish ([60d29e7](https://github.com/uvop/memorix/commit/60d29e71883c6a652261f40ff9983f7951b70bf3))
+* Try fix publish ([3b89b1a](https://github.com/uvop/memorix/commit/3b89b1a2b26f7fe5751d5cb0f61ac0f7068754c2))
+* try publish again ([0949d83](https://github.com/uvop/memorix/commit/0949d835551e6258f39baf5a2ebcf34b295efe68))
+* try publish again ([f33f417](https://github.com/uvop/memorix/commit/f33f4176c9f4286a085148024d0d9af63c104fd8))
+* try publish again ([7a05d3a](https://github.com/uvop/memorix/commit/7a05d3abce4989d056babfd46dec41dbdfef02e6))
+* Try redeploy ([428724c](https://github.com/uvop/memorix/commit/428724c3929a506c0ac2483e6752b462a68181af))
+
 # [1.0.0-next.7](https://github.com/uvop/memorix/compare/v1.0.0-next.6...v1.0.0-next.7) (2022-08-26)